SNOW-1062499 Remove old QueryStatus, replace it with QueryStatusV2
